Update OpenAI client integration and add title generation feature. Remove deprecated API code and refactor chat handling to streamline message processing. Enhance frontend configuration with proxy settings for API calls.

+++ backend/models/request.go
@@ -4,7 +4,7 @@ package models
 type ChatCompletionRequest struct {
 	Model       string        `json:"model"`
 	Messages    []ChatMessage `json:"messages"`
-	Temperature float64       `json:"temperature,omitempty"`
+	Temperature float32       `json:"temperature,omitempty"`
 	Stream      bool          `json:"stream,omitempty"`
 	MaxTokens   int           `json:"max_tokens,omitempty"`
 }
